update publish ui, created a publish activity
